requirements.txt
fastapi
uvicorn
jinja2
openai
requests
python-dotenv
python-multipart
beautifulsoup4
lxml
